Add files via upload
Car Price Prediction Project:
With the covid 19 impact in the market, we have seen a lot of changes in the car market. Now some 
cars are in demand, making them costly and some are not in demand, hence cheaper. One 
of our clients works with small traders, who sell used cars. With the change in the market due to covid 
19 impact, our client is facing problems with their previous car price valuation machine learning 
models. So, they are looking for new machine learning models from new data. We have to make 
a car price valuation model